Split permissions into separate files
